bank/pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = "Codebank core bank api"
authors = ["pauloreis7 <paulosilvadosreis2057@gmail.com>"]
license = "MIT"

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.2"
SQLAlchemy = "^1.4.40"
asyncpg = "^0.26.0"
alembic = "^1.8.1"
python-dotenv = "^0.20.0"
grpcio = "^1.48.0"
grpcio-tools = "^1.48.0"
confluent-kafka = "^1.9.2"
kafka = "^1.3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.14.5"
black = "^22.6.0"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
commitizen = "^2.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
